requirements.txt
fastapi
uvicorn
python-multipart
websocket-client
